+++ src/components/POLine/hooks/useIsFundsRestrictedByLocationIds/index.js
@@ -1,0 +1,1 @@
+export { useIsFundsRestrictedByLocationIds } from './useIsFundsRestrictedByLocationIds';